Clean up and fixes
